+++ spec/factories.rb
@@ -39,6 +39,10 @@ FactoryGirl.define do
     trait :with_project do
       after(:build) { |object| object.project = create(:project, users: [object.requested_by]) }
     end
+
+    trait :with_activity do
+      after(:create) { |object| create(:activity, subject: object, user: object.requested_by, project: object.project) }
+    end
   end
 
   factory :changeset do |c|