+++ spec/javascripts/models/note_spec.js
@@ -46,6 +46,10 @@ describe("Note", function() {
       I18n.t.withArgs('foo_bar').returns('Foo bar');
     });
 
+    afterEach(function() {
+      sinon.restore(I18n);
+    });
+
     it("returns the translated attribute name", function() {
       expect(this.note.humanAttributeName('foo_bar')).toEqual('Foo bar');
     });